Restructure gRPC protos and services
Split and rename gRPC proto contracts and add dedicated services
