--- electrostoreAPI/Grpc/Protos/iaToApi.proto
@@ -1,32 +1,0 @@
-syntax = "proto3";
-
-option csharp_namespace = "ElectrostoreAPI.Grpc";
-
-package iaToApi;
-
-service IAToAPIGrpc {
-  rpc StreamTrainingImages (StreamTrainingImagesRequest) returns (stream TrainingImage);
-
-  rpc GetConfig (IAGetConfigRequest) returns (IAGetConfigReply);
-}
-
-// ---- Training images ----
-
-message StreamTrainingImagesRequest {
-  repeated string existing_filenames = 1;
-}
-
-message TrainingImage {
-  string label    = 1;
-  string filename = 2;
-  bytes  data     = 3;
-}
-
-// ---- Config ----
-
-message IAGetConfigRequest {}
-
-message IAGetConfigReply {
-  bool            demo_mode                = 1;
-  repeated string allowed_image_extensions = 2;
-}
