--- electrostoreAPI/Grpc/Protos/workerToApi.proto
@@ -1,79 +1,0 @@
-syntax = "proto3";
-
-option csharp_namespace = "ElectrostoreAPI.Grpc";
-
-package workerToApi;
-
-service WORKERToAPIGrpc {
-  rpc UpdateCommandStatus    (UpdateCommandStatusRequest)    returns (UpdateCommandStatusReply);
-  rpc AddCommandHistory      (AddCommandHistoryRequest)      returns (AddCommandHistoryReply);
-  rpc UpdateStoreMqttStatus  (UpdateStoreMqttStatusRequest)  returns (UpdateStoreMqttStatusReply);
-  rpc UpdateIaStatus         (UpdateIaStatusRequest)         returns (UpdateIaStatusReply);
-  rpc GetConfig              (WORKERGetConfigRequest)         returns (WORKERGetConfigReply);
-}
-
-// ---- Command status update ----
-
-message UpdateCommandStatusRequest {
-  int32  id_command     = 1;
-  string status_command = 2;
-  string date_livraison = 3;
-}
-
-message UpdateCommandStatusReply {
-  bool success = 1;
-}
-
-// ---- Command history ----
-
-message AddCommandHistoryRequest {
-  int32  id_command     = 1;
-  string status         = 2;
-  string tracking_number = 3;
-  string carrier        = 4;
-  string tracking_event = 5;
-  string event_at       = 6; // ISO 8601
-}
-
-message AddCommandHistoryReply {
-  bool success              = 1;
-  int32 id_command_history  = 2;
-}
-
-// ---- Store MQTT status ----
-
-message UpdateStoreMqttStatusRequest {
-  string mqtt_name_store       = 1;
-  bool   is_mqtt_connected     = 2;
-}
-
-message UpdateStoreMqttStatusReply {
-  bool success      = 1;
-  int32 store_count = 2; // nombre de stores mis à jour
-}
-
-// ---- IA status ----
-
-message UpdateIaStatusRequest {
-  int32  id_ia        = 1;
-  string action       = 2; // training_started | training_completed | training_failed
-  int32  requested_by = 3;
-  string message      = 4;
-  float  accuracy     = 5;
-  float  val_accuracy = 6;
-  float  loss         = 7;
-  float  val_loss     = 8;
-  int32  epoch        = 9;
-}
-
-message UpdateIaStatusReply {
-  bool success = 1;
-}
-
-// ---- Config ----
-
-message WORKERGetConfigRequest {}
-
-message WORKERGetConfigReply {
-  bool demo_mode = 1;
-}
