--- electrostoreCRON/Grpc/Protos/cronToApi.proto
@@ -1,45 +1,0 @@
-syntax = "proto3";
-
-option csharp_namespace = "ElectrostoreCRON.Grpc";
-
-package cronToApi;
-
-service CRONToAPIGrpc {
-  rpc GetEnabledCronJobs (GetCronJobsRequest)     returns (GetCronJobsReply);
-  rpc UpdateCronJobRun   (UpdateCronJobRunRequest) returns (UpdateCronJobRunReply);
-  rpc GetConfig          (CRONGetConfigRequest)    returns (CRONGetConfigReply);
-}
-
-// ---- CronJobs ----
-
-message GetCronJobsRequest {}
-
-message CronJobItem {
-  int32  id_cronjob      = 1;
-  string name_cronjob    = 2;
-  string cron_expression = 3;
-  string action_cronjob  = 4;
-  string params_cronjob  = 5;
-}
-
-message GetCronJobsReply {
-  repeated CronJobItem jobs = 1;
-}
-
-message UpdateCronJobRunRequest {
-  int32  id_cronjob  = 1;
-  string last_run_at = 2; // ISO 8601
-  string next_run_at = 3; // ISO 8601
-}
-
-message UpdateCronJobRunReply {
-  bool success = 1;
-}
-
-// ---- Config ----
-
-message CRONGetConfigRequest {}
-
-message CRONGetConfigReply {
-  bool demo_mode = 1;
-}
