+++ electrostoreIA/Services/ConfigCacheService/ConfigCacheService.cs
@@ -5,13 +5,13 @@ namespace ElectrostoreIA.Services.ConfigCacheService;
 
 public class ConfigCacheService : IHostedService, IConfigCacheService
 {
-    private readonly IAToAPIGrpc.IAToAPIGrpcClient _client;
+    private readonly ConfigGrpc.ConfigGrpcClient _client;
     private readonly ILogger<ConfigCacheService> _logger;
 
     public bool DemoMode { get; private set; }
     public IEnumerable<string> AllowedImageExtensions { get; private set; } = [];
 
-    public ConfigCacheService(IAToAPIGrpc.IAToAPIGrpcClient client, ILogger<ConfigCacheService> logger)
+    public ConfigCacheService(ConfigGrpc.ConfigGrpcClient client, ILogger<ConfigCacheService> logger)
     {
         _client = client;
         _logger = logger;
