+++ electrostoreNOTIF/Services/ConfigCacheService/ConfigCacheService.cs
@@ -5,12 +5,12 @@ namespace ElectrostoreNOTIF.Services.ConfigCacheService;
 
 public class ConfigCacheService : IHostedService, IConfigCacheService
 {
-    private readonly NOTIFToAPIGrpc.NOTIFToAPIGrpcClient _client;
+    private readonly ConfigGrpc.ConfigGrpcClient _client;
     private readonly ILogger<ConfigCacheService> _logger;
 
     public bool DemoMode { get; private set; }
 
-    public ConfigCacheService(NOTIFToAPIGrpc.NOTIFToAPIGrpcClient client, ILogger<ConfigCacheService> logger)
+    public ConfigCacheService(ConfigGrpc.ConfigGrpcClient client, ILogger<ConfigCacheService> logger)
     {
         _client = client;
         _logger = logger;
