+++ electrostoreWORKER/Grpc/Protos/commands.proto
@@ -1,0 +1,38 @@
+syntax = "proto3";
+
+option csharp_namespace = "ElectrostoreWORKER.Grpc";
+
+package commands;
+
+service CommandsGrpc {
+  rpc UpdateCommandStatus    (UpdateCommandStatusRequest)    returns (UpdateCommandStatusReply);
+  rpc AddCommandHistory      (AddCommandHistoryRequest)      returns (AddCommandHistoryReply);
+}
+
+// ---- Command status update ----
+
+message UpdateCommandStatusRequest {
+  int32  id_command     = 1;
+  string status_command = 2;
+  string date_livraison = 3;
+}
+
+message UpdateCommandStatusReply {
+  bool success = 1;
+}
+
+// ---- Command history ----
+
+message AddCommandHistoryRequest {
+  int32  id_command     = 1;
+  string status         = 2;
+  string tracking_number = 3;
+  string carrier        = 4;
+  string tracking_event = 5;
+  string event_at       = 6; // ISO 8601
+}
+
+message AddCommandHistoryReply {
+  bool success              = 1;
+  int32 id_command_history  = 2;
+}
