+++ electrostoreWORKER/Program.cs
@@ -27,7 +27,22 @@ public partial class Program
         }
 
         // gRPC client to the API
-        builder.Services.AddGrpcClient<WORKERToAPIGrpc.WORKERToAPIGrpcClient>(options =>
+        builder.Services.AddGrpcClient<CommandsGrpc.CommandsGrpcClient>(options =>
+        {
+            options.Address = new Uri(
+                builder.Configuration["ApiServiceGrpcUrl"] ?? "http://electrostoreAPI:5001");
+        });
+        builder.Services.AddGrpcClient<ConfigGrpc.ConfigGrpcClient>(options =>
+        {
+            options.Address = new Uri(
+                builder.Configuration["ApiServiceGrpcUrl"] ?? "http://electrostoreAPI:5001");
+        });
+        builder.Services.AddGrpcClient<IaTrainingGrpc.IaTrainingGrpcClient>(options =>
+        {
+            options.Address = new Uri(
+                builder.Configuration["ApiServiceGrpcUrl"] ?? "http://electrostoreAPI:5001");
+        });
+        builder.Services.AddGrpcClient<StoresMqttGrpc.StoresMqttGrpcClient>(options =>
         {
             options.Address = new Uri(
                 builder.Configuration["ApiServiceGrpcUrl"] ?? "http://electrostoreAPI:5001");
